+++ util-read.R
@@ -13,7 +13,7 @@
 ##
 ## Copyright 2016-2018 by Claus Hunsen <hunsen@fim.uni-passau.de>
 ## Copyright 2017 by Raphael Nömmer <noemmer@fim.uni-passau.de>
-## Copyright 2017 by Christian Hechtl <hechtl@fim.uni-passau.de>
+## Copyright 2017-2018 by Christian Hechtl <hechtl@fim.uni-passau.de>
 ## Copyright 2017 by Felix Prasse <prassefe@fim.uni-passau.de>
 ## Copyright 2017 by Thomas Bock <bockthom@fim.uni-passau.de>
 ## All Rights Reserved.
@@ -290,9 +290,10 @@ read.authors = function(data.path) {
 ## / / / / / / / / / / / / / / / / / / / / / / / / / / / / / / / / / / / / /
 ## PaStA data --------------------------------------------------------------
 
-#' Read and parse the pasta data from the 'similar-mailbox' file.
+#' Read and parse the pasta data from the 'mbox-result' file.
 #' The form in the file is : <message-id> <possibly another message.id> ... => commit.hash commit.hash2 ....
-#' The parsed form is a data frame with message IDs as keys and commit hashes as values.
+#' The parsed form is a data frame with message IDs as keys, commit hashes as values, and a revision set id.
+#' If the message ID does not get mapped to a commit hash, the value for the commit hash is \code{NA}.
 #'
 #' @param data.path the path to the pasta data
 #'
@@ -303,7 +304,7 @@ read.pasta = function(data.path) {
     KEY.SEPERATOR = " "
 
     ## get file name of pasta data
-    filepath = file.path(data.path, "similar-mailbox")
+    filepath = file.path(data.path, "mbox-result")
 
     ## read data from disk [can be empty]
     lines = suppressWarnings(try(readLines(filepath), silent = TRUE))
@@ -315,13 +316,13 @@ read.pasta = function(data.path) {
         return(data.frame())
     }
 
-    result.list = parallel::mclapply(lines, function(line) {
+    result.list = parallel::mcmapply(lines, seq_along(lines), SIMPLIFY = FALSE, FUN = function(line, line.id) {
         #line = lines[i]
         if ( nchar(line) == 0 ) {
             return(NULL)
         }
 
-        if (!grepl(SEPERATOR, line)) {
+        if (!grepl("<", line)) {
             logging::logwarn("Faulty line: %s", line)
             return(NULL)
         }
@@ -330,16 +331,21 @@ read.pasta = function(data.path) {
         # 2) split keys
         # 3) split values
         # 4) insert all key-value pairs by iteration (works also if there is only one key)
-        line.split = unlist(strsplit(line, SEPERATOR))
-        keys = line.split[1]
-        values = line.split[2]
-        keys.split = unlist(strsplit(keys, KEY.SEPERATOR))
-        values.split = unlist(strsplit(values, KEY.SEPERATOR))
+        if (grepl(SEPERATOR, line)) {
+            line.split = unlist(strsplit(line, SEPERATOR))
+            keys = line.split[1]
+            values = line.split[2]
+            keys.split = unlist(strsplit(keys, KEY.SEPERATOR))
+            values.split = unlist(strsplit(values, KEY.SEPERATOR))
+        } else {
+            keys.split = unlist(strsplit(line, KEY.SEPERATOR))
+            values.split = NA
+        }
 
         # Transform data to data.frame
-        #df = data.frame(message.id = keys.split, commit.hash = values.split)
         df = merge(keys.split, values.split)
         colnames(df) = c("message.id", "commit.hash")
+        df["revision.set.id"] = sprintf("<revision-set-%s>", line.id)
         return(df)
     })
     result.df = plyr::rbind.fill(result.list)